Apply more suggestions from dey4ss

+++ src/lib/scheduler/node_queue_scheduler.cpp
@@ -53,18 +53,17 @@ void NodeQueueScheduler::begin() {
   // NUM_GROUPS_MIN_FACTOR=0.1 and 128 workers, we would not group task lists with less than 25 tasks (2 * 128 * 0.1).
   _min_task_count_for_regrouping =
       std::max(size_t{2 * MIN_GROUP_COUNT},
-               static_cast<size_t>(2.0f * static_cast<float>(_worker_count) * NUM_GROUPS_MIN_FACTOR));
+               static_cast<size_t>(2.0 * static_cast<double>(_worker_count) * NUM_GROUPS_MIN_FACTOR));
 
   // Starting from a queue load of `_queue_load_capped`, we use the minimal group count.
   _queue_load_capped = static_cast<double>(std::max(size_t{8}, UPPER_LIMIT_QUEUE_SIZE_FACTOR * _worker_count));
 
   _min_group_count =
       std::max(static_cast<double>(MIN_GROUP_COUNT), NUM_GROUPS_MIN_FACTOR * static_cast<float>(_worker_count));
-  _max_group_count = NUM_GROUPS_MAX_FACTOR * static_cast<float>(_worker_count);
+  _max_group_count = NUM_GROUPS_MAX_FACTOR * static_cast<double>(_worker_count);
 
   // We later
   _step_size = (_max_group_count - _min_group_count) / _queue_load_capped;
-  // std::cout << ((_step_size * _queue_load_capped) == (_max_group_count - _min_group_count)) << " <<<<<\n";
   Assert((_step_size * _queue_load_capped) == (_max_group_count - _min_group_count), "...");
 
   // For every task list of size >= `_regrouping_upper_limit`, we use the minimal value for grouping. For 64 workers, a
